Updated the page
Updated the page and added files for the login function. The login works through cookies. I added cookie options and cookie expiration dates. I added the option for the code to check whether the user password or email is incorrect and return an appropriate result.

As of right now, we don't have a logout button. After successfully logging in, the page just redirects you to the home page. This is where we start building the main page itself.

ps. also some smaller changes were made to duplicate code sections.